requirements.txt
Flask-Login==0.5.0
Flask-Bcrypt==0.7.1
stripe==2.55.1
python-dotenv==0.19.2
